+++ Rain.Wave/IWave.cs
@@ -4,6 +4,6 @@ namespace Rain.Wave
 {
 	public interface IWave
 	{
-		float Probe(float time);
+		double Probe(double time);
 	}
 }
